Files for timestep determination
Function file that uses the NSimulateMotor function to input various k values as the input and having speed and processivity as the output.
TimestepDeterminationExperiment is a file where you input the timesteps using the previously mentioned function to generate data. More work still needs to be done as the expected outputs are not being produced.